Refactor logging, add `binexport_core` libray
This allows to share more code with the new Binary Ninja exporter (to be
added later).

Drive-by: .gitignore and ClangFormat settings
