+++ CMakeLists.txt
@@ -20,7 +20,6 @@ include(GoogleTest) # Needs to come before changing module path
 list(APPEND CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR})
 include(copts.cmake)
 include(util.cmake)
-
 set(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
 
 # Use a copy of a subset of Boost that we ship
@@ -110,6 +109,8 @@ add_library(binexport_shared STATIC
   util/format.h
   util/process.cc
   util/process.h
+  util/logging.cc
+  util/logging.h
   util/status_macros.h
   util/statusor.h
   util/timer.h
@@ -174,26 +175,8 @@ gtest_discover_tests(binexport_test)
 # binexport2dump tool
 add_subdirectory(tools)
 
-# Code shared with the BinDiff plugin
-add_ida_library(binexport_plugin_shared STATIC
-  hash.cc
-  hash.h
-  ida/digest.cc
-  ida/digest.h
-  ida/log.cc
-  ida/log.h
-  ida/ui.cc
-  ida/ui.h
-  ida/util.cc
-  ida/util.h
-)
-ida_target_link_libraries(binexport_plugin_shared PUBLIC
-  absl::time
-  binexport_base
-)
-
-set(binexport_plugin_name binexport${binexport_VERSION_MAJOR})
-add_ida_plugin(${binexport_plugin_name}
+# Non-plugin code shared with BinDiff/the Binary Ninja plugin
+add_library(binexport_core
   address_references.cc
   address_references.h
   base_types.cc
@@ -206,16 +189,6 @@ add_ida_plugin(${binexport_plugin_name}
   call_graph.h
   comment.cc
   comment.h
-  database/initialize_constraints_postgresql_sql.h
-  database/initialize_indices_postgresql_sql.h
-  database/initialize_tables_postgresql_sql.h
-  database/maintenance_postgresql_sql.h
-  database/postgresql.cc
-  database/postgresql.h
-  database/postgresql_writer.cc
-  database/postgresql_writer.h
-  database/query_builder.cc
-  database/query_builder.h
   dump_writer.cc
   dump_writer.h
   edge.cc
@@ -230,6 +203,62 @@ add_ida_plugin(${binexport_plugin_name}
   flow_graph.h
   function.cc
   function.h
+  instruction.cc
+  instruction.h
+  library_manager.cc
+  library_manager.h
+  operand.cc
+  operand.h
+  statistics_writer.cc
+  statistics_writer.h
+  type_system.cc
+  type_system.h
+  version.h
+  ${CMAKE_CURRENT_BINARY_DIR}/version.cc
+  virtual_memory.cc
+  virtual_memory.h
+  x86_nop.cc
+  x86_nop.h
+)
+target_link_libraries(binexport_core PUBLIC
+  absl::bad_optional_access
+  absl::flat_hash_map
+  absl::flat_hash_set
+  absl::hash
+  absl::node_hash_map
+  absl::node_hash_set
+  absl::optional
+  absl::strings
+  absl::time
+  binexport_shared
+)
+
+# Code shared with the BinDiff plugin
+add_ida_library(binexport_plugin_shared STATIC
+  hash.cc
+  hash.h
+  ida/digest.cc
+  ida/digest.h
+  ida/log_handler.cc
+  ida/log_handler.h
+)
+ida_target_link_libraries(binexport_plugin_shared PUBLIC
+  absl::time
+  binexport_base
+)
+
+set(binexport_plugin_name binexport${binexport_VERSION_MAJOR})
+add_ida_plugin(${binexport_plugin_name}
+  database/initialize_constraints_postgresql_sql.h
+  database/initialize_indices_postgresql_sql.h
+  database/initialize_tables_postgresql_sql.h
+  database/maintenance_postgresql_sql.h
+  database/postgresql.cc
+  database/postgresql.h
+  database/postgresql_writer.cc
+  database/postgresql_writer.h
+  database/query_builder.cc
+  database/query_builder.h
   ida/arm.cc
   ida/arm.h
   ida/dalvik.cc
@@ -249,22 +278,8 @@ add_ida_plugin(${binexport_plugin_name}
   ida/ppc.h
   ida/types_container.cc
   ida/types_container.h
-  instruction.cc
-  instruction.h
-  library_manager.cc
-  library_manager.h
-  operand.cc
-  operand.h
-  statistics_writer.cc
-  statistics_writer.h
-  type_system.cc
-  type_system.h
-  version.h
-  ${CMAKE_CURRENT_BINARY_DIR}/version.cc
-  virtual_memory.cc
-  virtual_memory.h
-  x86_nop.cc
-  x86_nop.h
+  ida/ui.cc
+  ida/ui.h
 )
 if(WIN32)
   list(APPEND binexport_libraries crypt32.lib
@@ -283,7 +298,7 @@ ida_target_link_libraries(${binexport_plugin_name}
   absl::optional
   absl::strings
   absl::time
-  binexport_shared
+  binexport_core
   binexport_plugin_shared
   ${PostgreSQL_LIBRARIES}
   # OpenSSL must come after PostgreSQL
@@ -298,7 +313,7 @@ ida_install(TARGETS ${binexport_plugin_name}
             RUNTIME DESTINATION binexport-prefix
             LIBRARY DESTINATION binexport-prefix)
 
-# Experimental BinExport reader library
+# BinExport reader library
 add_library(binexport_reader STATIC
   reader/call_graph.cc
   reader/call_graph.h
