+++ ida/log_handler.h
@@ -1,0 +1,27 @@
+// Copyright 2011-2019 Google LLC. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef IDA_LOG_HANDLER_H_
+#define IDA_LOG_HANDLER_H_
+
+#include "base/logging.h"
+
+namespace security::binexport {
+
+void IdaLogHandler(LogLevel level, const char* filename, int line,
+                   const std::string& message);
+
+}  // namespace security::binexport
+
+#endif  // IDA_LOG_HANDLER_H_
